requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.16
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
decorator==4.4.2
diffusers==0.30.3
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
eval_type_backport==0.2.0
fairscale==0.4.13
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
googletrans==3.0.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.25.1
hyperframe==5.2.0
idna==2.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
np==1.0.2
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
peft==0.13.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.11
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
rich==13.8.1
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.38.6
sympy==1.13.3
tasklib==2.5.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tokenizers==0.20.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.102
ultralytics-thop==2.0.8
urllib3==2.2.3
urwid==2.6.15
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2